Move results components
