+++ frontend/components/emission-results/all-category-emissions-summary.tsx
